separate different db interface nodes into multiple files (for version-2)
